requirements.txt
pandas
numpy
openai
python-dotenv
pytube
gradio_client
wandb
pathlib
langchain
